api/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
redis
qdrant-client
python-dotenv